@@ -47,12 +47,12 @@ class RelaxWorkChain(WorkChain):
             required=False,
         )
         spec.input(
-            'perform_static',
+            'relax.perform_static',
             valid_type=get_data_class('core.bool'),
             required=False,
             default=lambda: get_data_node('core.bool', True),
             help="""
-            If True, perform static calculation after relaxation.
+            If True, perform static calculation after relaxation to obtain a more sane electronic structure etc.
             """,
         )
         spec.input(
@@ -234,7 +234,7 @@ class RelaxWorkChain(WorkChain):
                 ),
                 cls.store_relaxed,
             ),
-            if_(cls.perform_final_static)(
+            if_(cls.perform_static)(
                 cls.init_relaxed,
                 cls.init_next_workchain,
                 cls.run_next_workchain,
@@ -260,6 +260,7 @@ class RelaxWorkChain(WorkChain):
         self.ctx.is_converged = False
         self.ctx.current_site_magnetization = None
         self.ctx.relax = False
+        self.ctx.perform_static = False
         self.ctx.iteration = 0
         self.ctx.workchains = []
         self.ctx.inputs = AttributeDict()
@@ -303,6 +304,7 @@ class RelaxWorkChain(WorkChain):
 
         # Need to save the parameter that controls if relaxation should be performed
         self.ctx.relax = parameters.relax.perform
+        self.ctx.perform_static = parameters.relax.perform_static
         if not parameters.relax.perform:
             # Make sure we do not expose the relax namespace in the input parameters (
             # basically setting no code tags related to relaxation unless user overrides)
@@ -384,15 +386,17 @@ class RelaxWorkChain(WorkChain):
             return self.exit_codes.ERROR_NO_CALLED_WORKCHAIN  # pylint: disable=no-member
 
         # Inherit exit status from last workchain (supposed to be
-        # successfull)
+        # successful)
         next_workchain_exit_status = workchain.exit_status
         next_workchain_exit_message = workchain.exit_message
         if not next_workchain_exit_status:
             self.ctx.exit_code = self.exit_codes.NO_ERROR  # pylint: disable=no-member
         else:
             self.ctx.exit_code = compose_exit_code(next_workchain_exit_status, next_workchain_exit_message)
-            self.report('The called {}<{}> returned a non-zero exit status. '
-                        'The exit status {} is inherited'.format(workchain.__class__.__name__, workchain.pk, self.ctx.exit_code))
+            self.report(
+                f'The called {workchain.__class__.__name__}<{workchain.pk}> returned a non-zero exit status. '
+                f'The exit status {self.ctx.exit_code} is inherited'
+            )
             # Make sure at the very minimum we attach the misc node (if it exists) that contains notifications and other
             # quantities that can be salvaged
             try:
@@ -412,9 +416,11 @@ class RelaxWorkChain(WorkChain):
         workchain = self.ctx.workchains[-1]
         # Double check presence of structure
         if 'structure' not in workchain.outputs:
-            self.report('The {}<{}> for the relaxation run did not have an '
-                        'output structure and most likely failed. However, '
-                        'its exit status was zero.'.format(workchain.__class__.__name__, workchain.pk))
+            self.report(
+                f'The {workchain.__class__.__name__}<{workchain.pk}> for the relaxation run did not have an '
+                'output structure and most likely failed. However, '
+                'its exit status was zero.'
+            )
             return self.exit_codes.ERROR_NO_RELAXED_STRUCTURE  # pylint: disable=no-member
 
         # Because the workchain may have been through multiple restarts of the underlying VASP calculation
@@ -464,13 +470,17 @@ class RelaxWorkChain(WorkChain):
         """Check the difference in cell shape before / after the last iteration against a tolerance."""
         lengths_converged = bool(delta.cell_lengths.max() <= self.ctx.inputs.parameters.relax.convergence_shape_lengths)
         if not lengths_converged:
-            self.report('cell lengths changed by max {}, tolerance is {}'.format(
-                delta.cell_lengths.max(), self.ctx.inputs.parameters.relax.convergence_shape_lengths))
+            self.report(
+                f'cell lengths changed by max {delta.cell_lengths.max()}, '
+                f'tolerance is {self.ctx.inputs.parameters.relax.convergence_shape_lengths}'
+            )
 
         angles_converged = bool(delta.cell_angles.max() <= self.ctx.inputs.parameters.relax.convergence_shape_angles)
         if not angles_converged:
-            self.report('cell angles changed by max {}, tolerance is {}'.format(delta.cell_angles.max(),
-                                                                                self.ctx.inputs.parameters.relax.convergence_shape_angles))
+            self.report(
+                f'cell angles changed by max {delta.cell_angles.max()}, '
+                f'tolerance is {self.ctx.inputs.parameters.relax.convergence_shape_angles}'
+            )
 
         return bool(lengths_converged and angles_converged)
 
@@ -478,25 +488,33 @@ class RelaxWorkChain(WorkChain):
         """Check the convergence of the volume, given a cutoff."""
         volume_converged = bool(delta.volume <= self.ctx.inputs.parameters.relax.convergence_volume)
         if not volume_converged:
-            self.report('cell volume changed by {}, tolerance is {}'.format(delta.volume,
-                                                                            self.ctx.inputs.parameters.relax.convergence_volume))
+            self.report(
+                f'cell volume changed by {delta.volume}, '
+                f'tolerance is {self.ctx.inputs.parameters.relax.convergence_volume}'
+            )
         return volume_converged
 
     def check_positions_convergence(self, delta):
         """Check the convergence of the atomic positions, given a cutoff."""
         try:
-            positions_converged = bool(np.nanmax(delta.pos_lengths) <= self.ctx.inputs.parameters.relax.convergence_positions)
+            positions_converged = bool(
+                np.nanmax(delta.pos_lengths) <= self.ctx.inputs.parameters.relax.convergence_positions
+            )
         except RuntimeWarning:
             # Here we encountered the case of having one atom centered at the origin, so
             # we do not know if it is converged, so settings it to False
-            self.report('there is NaN entries in the relative comparison for '
-                        'the positions during relaxation, assuming position is not converged')
+            self.report(
+                'there is NaN entries in the relative comparison for '
+                'the positions during relaxation, assuming position is not converged'
+            )
             positions_converged = False
 
         if not positions_converged:
             try:
-                self.report('max site position change is {}, tolerance is {}'.format(np.nanmax(
-                    delta.pos_lengths), self.ctx.inputs.parameters.relax.convergence_positions))
+                self.report(
+                    f'max site position change is {np.nanmax(delta.pos_lengths)}, '
+                    f'tolerance is {self.ctx.inputs.parameters.relax.convergence_positions}'
+                )
             except RuntimeWarning:
                 pass
 
@@ -508,8 +526,9 @@ class RelaxWorkChain(WorkChain):
 
         relaxed_structure = workchain.outputs.structure
         if self._verbose:
-            self.report("attaching the node {}<{}> as '{}'".format(relaxed_structure.__class__.__name__, relaxed_structure.pk,
-                                                                   'relax.structure'))
+            self.report(
+                f"attaching the node {relaxed_structure.__class__.__name__}<{relaxed_structure.pk}> as 'relax.structure'"
+            )
         self.out('relax.structure', relaxed_structure)
 
     def results(self):
@@ -525,6 +544,6 @@ class RelaxWorkChain(WorkChain):
         """Check if a relaxation is to be performed."""
         return self.ctx.relax
 
-    def perform_final_static(self):
+    def perform_static(self):
         """Check if the final static calculation is to be performed."""
-        return self.inputs.perform_static
+        return self.ctx.perform_static
