@@ -1,3 +1,3 @@
-"""AiiDA Plugin for running VASP calculations."""
+"""AiiDA plugin for running and managing VASP calculations."""
 
-__version__ = '2.1.1'
+__version__ = '3.0.1'
