@@ -5,8 +5,8 @@ import os
 from pathlib import Path
 from typing import Union
 
-from aiida import orm
 from aiida.common.exceptions import InputValidationError
+from aiida.plugins import DataFactory
 
 from aiida_vasp.calcs.vasp import VaspCalculation, ordered_unique_symbols
 from aiida_vasp.parsers.content_parsers.poscar import PoscarParser
@@ -270,7 +270,9 @@ class VaspNEBCalculation(VaspCalculation):
             store = True
 
         try:
-            additional_retrieve_list = self.inputs.settings.base.attributes.get('PER_IMAGE_ADDITIONAL_RETRIEVE_LIST', default=[])
+            additional_retrieve_list = self.inputs.settings.base.attributes.get(
+                'PER_IMAGE_ADDITIONAL_RETRIEVE_LIST', default=[]
+            )
         except AttributeError:
             additional_retrieve_list = []
         try:
@@ -281,11 +283,15 @@ class VaspNEBCalculation(VaspCalculation):
 
         if store:
             calcinfo.retrieve_list.extend(
-                image_folder_paths(image_folders, set(self._PER_IMAGE_ALWAYS_RETRIEVE_LIST + additional_retrieve_list)))
+                image_folder_paths(image_folders, set(self._PER_IMAGE_ALWAYS_RETRIEVE_LIST + additional_retrieve_list))
+            )
             calcinfo.retrieve_temporary_list.extend(image_folder_paths(image_folders, additional_retrieve_temp_list))
         else:
             calcinfo.retrieve_temporary_list.extend(
-                image_folder_paths(image_folders, set(self._PER_IMAGE_ALWAYS_RETRIEVE_LIST + additional_retrieve_temp_list)))
+                image_folder_paths(
+                    image_folders, set(self._PER_IMAGE_ALWAYS_RETRIEVE_LIST + additional_retrieve_temp_list)
+                )
+            )
             calcinfo.retrieve_list.extend(image_folder_paths(image_folders, additional_retrieve_list))
 
         #self.logger.warning('Calcinfo: {}'.format(calcinfo))
@@ -358,7 +364,8 @@ class VaspNEBCalculation(VaspCalculation):
         super().verify_inputs()
         last_order = None
         last_num_sites = None
-        for structure in list(self.inputs.neb_images.values()) + [self.inputs.initial_structure, self.inputs.final_structure]:
+        for structure in list(self.inputs.neb_images.values()
+                              ) + [self.inputs.initial_structure, self.inputs.final_structure]:
             # Convert to StructureData from CifData on demand....
             if not hasattr(structure, 'get_pymatgen'):
                 structure = get_data_node('core.structure', ase=structure.get_ase())
@@ -398,9 +405,11 @@ def image_folder_paths(image_folders, retrieve_names):
     return retrieve_list
 
 
-def ensure_structure_data(structure: Union[orm.StructureData, orm.CifData]) -> orm.StructureData:
+def ensure_structure_data(
+    structure: Union[DataFactory('core.structure'), DataFactory('core.cif')]
+) -> DataFactory('core.structure'):
     """
-    Get the input structure as AiiDA StructureData.
+        Get the input structure as AiiDA StructureData.
 
     This is required in order to support CifData as input as well.
     """
