@@ -55,18 +55,23 @@ def detect_old_style_groups():
         if count == 0:
             not_migrated.append(group_label)
     if any(not_migrated):
-        click.echo(
-            ("Some of the old style POTCAR family groups are not migrated. Please run command 'verdi data vasp-potcar migratefamilies.\n'",
-             f'The missing groups are: {not_migrated}.'))
+        click.echo((
+            "Some of the old style POTCAR family groups are not migrated. Please run command 'verdi data vasp-potcar migratefamilies.\n'",
+            f'The missing groups are: {not_migrated}.'
+        ))
 
 
 @potcar.command()
-@options.PATH(help='Path to a folder or archive containing the POTCAR files. '
-              'You can supply the archive that you downloaded from the VASP server. '
-              'The path does not need to be specified, if that is the case, the current path is used.')
+@options.PATH(
+    help='Path to a folder or archive containing the POTCAR files. '
+    'You can supply the archive that you downloaded from the VASP server. '
+    'The path does not need to be specified, if that is the case, the current path is used.'
+)
 @options.FAMILY_NAME()
 @options.DESCRIPTION(help='A description for the family.', callback=try_grab_description)
-@click.option('--stop-if-existing', is_flag=True, help='An option to abort when encountering a previously uploaded POTCAR file.')
+@click.option(
+    '--stop-if-existing', is_flag=True, help='An option to abort when encountering a previously uploaded POTCAR file.'
+)
 @options.DRY_RUN()
 @with_dbenv()
 def uploadfamily(path, name, description, stop_if_existing, dry_run):
@@ -74,11 +79,9 @@ def uploadfamily(path, name, description, stop_if_existing, dry_run):
 
     potcar_data_cls = get_data_class('vasp.potcar')
     with cli_spinner():
-        num_found, num_added, num_uploaded = potcar_data_cls.upload_potcar_family(path,
-                                                                                  name,
-                                                                                  description,
-                                                                                  stop_if_existing=stop_if_existing,
-                                                                                  dry_run=dry_run)
+        num_found, num_added, num_uploaded = potcar_data_cls.upload_potcar_family(
+            path, name, description, stop_if_existing=stop_if_existing, dry_run=dry_run
+        )
 
     click.echo(f'POTCAR files found: {num_found}. New files uploaded: {num_uploaded}, Added to Family: {num_added}')
     if dry_run:
@@ -86,7 +89,9 @@ def uploadfamily(path, name, description, stop_if_existing, dry_run):
 
 
 @potcar.command()
-@click.option('-e', '--element', multiple=True, help='Filter for families containing potentials for all given elements.')
+@click.option(
+    '-e', '--element', multiple=True, help='Filter for families containing potentials for all given elements.'
+)
 @click.option('-s', '--symbol', multiple=True, help='Filter for families containing potentials for all given symbols.')
 @click.option('-d', '--description', is_flag=True, help='Also show the description.')
 @with_dbenv()
