@@ -4,7 +4,6 @@ The ``POSCAR``/``CONTCAR`` parser interface.
 ------------------------------------
 Contains the parsing interfaces to ``parsevasp`` used to parse ``POSCAR``/``CONTCAR`` content.
 """
-# pylint: disable=no-self-use
 import numpy as np
 from parsevasp.poscar import Poscar, Site
 
@@ -55,7 +54,9 @@ class PoscarParser(BaseFileParser):
         """
 
         try:
-            self._content_parser = Poscar(file_handler=handler, prec=self._precision, conserve_order=True, logger=self._logger)
+            self._content_parser = Poscar(
+                file_handler=handler, prec=self._precision, conserve_order=True, logger=self._logger
+            )
         except SystemExit:
             self._logger.warning('Parsevasp exited abnormally.')
 
