@@ -90,8 +90,8 @@ def compare_doscar_content(result, pdos=True):
                              (4.165, 1.163e-01, 2.961e-02, 2.961e-02, 2.961e-02, 0., 0., 0., 0., 0.),
                              (6.066, -5.870e-17, -2.973e-17, -2.973e-17, -2.973e-17, 0., 0., 0., 0., 0.),
                              (7.967, 7.681e-02, 3.890e-02, 3.890e-02, 3.890e-02, 0., 0., 0., 0., 0.),
-                             (9.868, -1.988e-29, -1.936e-29, -1.923e-29, -1.944e-29, 0., 0., 0., 0., 0.),
-                             (11.769, 5.486e-02, 3.117e-02, 3.938e-02, 3.827e-02, 0., 0., 0., 0., 0.),
+                             (9.868, -2.042e-29, -1.886e-29, -1.898e-29, -1.932e-29, 0., 0., 0., 0., 0.),
+                             (11.769, 4.011e-02, 2.855e-02, 3.605e-02, 3.375e-02, 0., 0., 0., 0., 0.),
                              (13.67, 0.000e+00, 0.000e+00, 0.000e+00, 0.000e+00, 0., 0., 0., 0., 0.)],
                             [(-3.44, -6.543e-45, -3.040e-46, -3.040e-46, -3.040e-46, 0., 0., 0., 0., 0.),
                              (-1.539, 8.295e-03, 3.853e-04, 3.853e-04, 3.853e-04, 0., 0., 0., 0., 0.),
@@ -100,8 +100,8 @@ def compare_doscar_content(result, pdos=True):
                              (4.165, 1.163e-01, 2.961e-02, 2.961e-02, 2.961e-02, 0., 0., 0., 0., 0.),
                              (6.066, -5.870e-17, -2.973e-17, -2.973e-17, -2.973e-17, 0., 0., 0., 0., 0.),
                              (7.967, 7.681e-02, 3.890e-02, 3.890e-02, 3.890e-02, 0., 0., 0., 0., 0.),
-                             (9.868, -1.988e-29, -1.936e-29, -1.923e-29, -1.944e-29, 0., 0., 0., 0., 0.),
-                             (11.769, 5.486e-02, 3.117e-02, 3.938e-02, 3.827e-02, 0., 0., 0., 0., 0.),
+                             (9.868, -2.061e-29, -1.918e-29, -1.894e-29, -1.872e-29, 0., 0., 0., 0., 0.),
+                             (11.769, 7.493e-02, 3.673e-02, 4.614e-02, 3.762e-02, 0., 0., 0., 0., 0.),
                              (13.67, 0.000e+00, 0.000e+00, 0.000e+00, 0.000e+00, 0., 0., 0., 0., 0.)],
                             [(-3.44, -6.543e-45, -3.040e-46, -3.040e-46, -3.040e-46, 0., 0., 0., 0., 0.),
                              (-1.539, 8.295e-03, 3.853e-04, 3.853e-04, 3.853e-04, 0., 0., 0., 0., 0.),
@@ -110,11 +110,12 @@ def compare_doscar_content(result, pdos=True):
                              (4.165, 1.163e-01, 2.961e-02, 2.961e-02, 2.961e-02, 0., 0., 0., 0., 0.),
                              (6.066, -5.870e-17, -2.973e-17, -2.973e-17, -2.973e-17, 0., 0., 0., 0., 0.),
                              (7.967, 7.681e-02, 3.890e-02, 3.890e-02, 3.890e-02, 0., 0., 0., 0., 0.),
-                             (9.868, -1.988e-29, -1.936e-29, -1.923e-29, -1.944e-29, 0., 0., 0., 0., 0.),
-                             (11.769, 5.486e-02, 3.117e-02, 3.938e-02, 3.827e-02, 0., 0., 0., 0., 0.),
+                             (9.868, -2.089e-29, -1.874e-29, -1.898e-29, -1.865e-29, 0., 0., 0., 0., 0.),
+                             (11.769, 7.120e-02, 3.744e-02, 4.373e-02, 3.642e-02, 0., 0., 0., 0., 0.),
                              (13.67, 0.000e+00, 0.000e+00, 0.000e+00, 0.000e+00, 0., 0., 0., 0., 0.)]],
-                           dtype=[('energy', '<f8'), ('s', '<f8'), ('py', '<f8'), ('px', '<f8'), ('pz', '<f8'), ('dxy', '<f8'),
-                                  ('dyz', '<f8'), ('dz2', '<f8'), ('dxz', '<f8'), ('dx2-y2', '<f8')])
+                           dtype=[('energy', '<f8'), ('s', '<f8'), ('py', '<f8'), ('px', '<f8'), ('pz', '<f8'),
+                                  ('dxy', '<f8'), ('dyz', '<f8'), ('dz2', '<f8'), ('dxz', '<f8'), ('dx2-y2', '<f8')])
+
     result_total_dos = result['tdos']
     for item in total_dos.dtype.names:
         assert np.allclose(result_total_dos[item], total_dos[item])
