@@ -32,7 +32,10 @@ def test_parse_outcar_no_quantity(outcar_parser):
     assert outcar_parser.get_quantity('magnetization') is None
 
 
-@pytest.mark.parametrize(['outcar_parser'], [(['disp_details', 'OUTCAR', {'quantities_to_parse': ['symmetries']}],)], indirect=True)
+@pytest.mark.parametrize(['outcar_parser'], [(['disp_details', 'OUTCAR', {
+    'quantities_to_parse': ['symmetries']
+}],)],
+                         indirect=True)
 def test_parse_outcar_symmetry(outcar_parser, compare_symmetries):
     """Load a reference OUTCAR parser.
 
@@ -44,7 +47,10 @@ def test_parse_outcar_symmetry(outcar_parser, compare_symmetries):
     assert set(symmetry) == set(compare_symmetries)
 
 
-@pytest.mark.parametrize(['outcar_parser'], [(['disp_details', 'OUTCAR', {'quantities_to_parse': ['elastic_moduli']}],)], indirect=True)
+@pytest.mark.parametrize(['outcar_parser'], [(['disp_details', 'OUTCAR', {
+    'quantities_to_parse': ['elastic_moduli']
+}],)],
+                         indirect=True)
 def test_parse_outcar_elastic_moduli(outcar_parser):
     """Load a reference OUTCAR parser.
 
@@ -109,10 +115,12 @@ def test_parse_outcar_status_extended(outcar_parser, expected):
     assert status['contains_nelm_breach'] is expected[4]
 
 
-@pytest.mark.parametrize(['outcar_parser'], [(['magnetization', 'OUTCAR', {
-    'quantities_to_parse': ['magnetization', 'site_magnetization']
-}],)],
-                         indirect=True)
+@pytest.mark.parametrize(
+    ['outcar_parser'], [(['magnetization', 'OUTCAR', {
+        'quantities_to_parse': ['magnetization', 'site_magnetization']
+    }],)],
+    indirect=True
+)
 def test_parse_outcar_magnetizationr(fresh_aiida_env, outcar_parser):
     """Load a reference OUTCAR parser.
 
@@ -170,11 +178,13 @@ def test_parse_outcar_magnetizationr(fresh_aiida_env, outcar_parser):
     assert set(magnetization) == set(test)
 
 
-@pytest.mark.parametrize(['outcar_parser'],
-                         [(['magnetization_single', 'OUTCAR', {
-                             'quantities_to_parse': ['magnetization', 'site_magnetization']
-                         }],)],
-                         indirect=True)
+@pytest.mark.parametrize(
+    ['outcar_parser'],
+    [(['magnetization_single', 'OUTCAR', {
+        'quantities_to_parse': ['magnetization', 'site_magnetization']
+    }],)],
+    indirect=True
+)
 def test_parse_outcar_magnetization_single(fresh_aiida_env, outcar_parser):  # pylint: disable=invalid-name
     """Load a reference OUTCAR parser.
 
