@@ -39,12 +39,14 @@ def test_parse_poscar_write(poscar_parser, tmpdir):
     with open(temp_path, 'r', encoding='utf8') as handler:
         content = handler.readlines()
     ref_content = [
-        '# Compound: Al4. Old comment: Al\n', '  1.000000000000\n', '  4.040000000000   0.000000000000   0.000000000000\n',
-        '  0.000000000000   4.040000000000   0.000000000000\n', '  0.000000000000   0.000000000000   4.040000000000\n', 'Al\n', '    4\n',
-        'Direct\n', '  0.000000000000   0.000000000000   0.000000000000\n', '  0.000000000000   0.500000000000   0.500000000000\n',
-        '  0.500000000000   0.000000000000   0.500000000000\n', '  0.500000000000   0.500000000000   0.000000000000\n', '\n',
+        '# Compound: Al4. Old comment: Al\n', '  1.000000000000\n',
+        '  4.040000000000   0.000000000000   0.000000000000\n', '  0.000000000000   4.040000000000   0.000000000000\n',
+        '  0.000000000000   0.000000000000   4.040000000000\n', 'Al\n', '    4\n', 'Direct\n',
+        '  0.000000000000   0.000000000000   0.000000000000\n', '  0.000000000000   0.500000000000   0.500000000000\n',
+        '  0.500000000000   0.000000000000   0.500000000000\n', '  0.500000000000   0.500000000000   0.000000000000\n',
+        '\n', '  0.000000000000   0.000000000000   0.000000000000\n',
         '  0.000000000000   0.000000000000   0.000000000000\n', '  0.000000000000   0.000000000000   0.000000000000\n',
-        '  0.000000000000   0.000000000000   0.000000000000\n', '  0.000000000000   0.000000000000   0.000000000000\n'
+        '  0.000000000000   0.000000000000   0.000000000000\n'
     ]
     assert content == ref_content
 
