@@ -50,11 +50,12 @@ from .data import (
 from .environment import fresh_aiida_env
 
 __all__ = [
-    'fresh_aiida_env', 'localhost_dir', 'localhost', 'vasp_params', 'potentials', 'vasp_structure', 'vasp_kpoints', 'vasp_code',
-    'ref_incar', 'ref_incar_vasp2w90', 'vasp_chgcar', 'vasp_wavecar', 'ref_retrieved', 'vasp_calc_and_ref', 'vasp2w90_calc_and_ref',
-    'vasp2w90_calc', 'vasp_structure_poscar', 'temp_pot_folder', 'potcar_family', 'poscar_parser', 'doscar_parser', 'incar_parser',
-    'chgcar_parser', 'kpoints_parser', 'eigenval_parser', 'vasprun_parser', 'vasprun_parser_v621', 'outcar_parser', 'mock_vasp',
-    'mock_vasp_strict', 'wannier_params', 'wannier_projections', 'ref_win', 'phonondb_run', 'base_calc', 'vasp_calc', 'vasp_inputs',
-    'vasp2w90_inputs', 'run_vasp_process', 'stream_parser', 'compare_symmetries', 'neb_outcar_parser', 'vasp_neb_calc', 'vasp_neb_inputs',
+    'fresh_aiida_env', 'localhost_dir', 'localhost', 'vasp_params', 'potentials', 'vasp_structure', 'vasp_kpoints',
+    'vasp_code', 'ref_incar', 'ref_incar_vasp2w90', 'vasp_chgcar', 'vasp_wavecar', 'ref_retrieved', 'vasp_calc_and_ref',
+    'vasp2w90_calc_and_ref', 'vasp2w90_calc', 'vasp_structure_poscar', 'temp_pot_folder', 'potcar_family',
+    'poscar_parser', 'doscar_parser', 'incar_parser', 'chgcar_parser', 'kpoints_parser', 'eigenval_parser',
+    'vasprun_parser', 'vasprun_parser_v621', 'outcar_parser', 'mock_vasp', 'mock_vasp_strict', 'wannier_params',
+    'wannier_projections', 'ref_win', 'phonondb_run', 'base_calc', 'vasp_calc', 'vasp_inputs', 'vasp2w90_inputs',
+    'run_vasp_process', 'stream_parser', 'compare_symmetries', 'neb_outcar_parser', 'vasp_neb_calc', 'vasp_neb_inputs',
     'sandbox_folder'
 ]
