@@ -40,8 +40,9 @@ def test_nested_dict_update(init_dict1, init_dict2):
     dict1 = AttributeDict(init_dict1)
     dict2 = AttributeDict(init_dict2)
     update_nested_dict(dict1, dict2)
-    test = AttributeDict(
-        {'dct1': AttributeDict({
+    test = AttributeDict({
+        'dct1':
+        AttributeDict({
             'test': 2.0,
             'test2': 'string2',
             'test3': [2.0],
@@ -49,7 +50,8 @@ def test_nested_dict_update(init_dict1, init_dict2):
                 'key1': 'value1',
                 'key2': 'value2'
             }
-        })})
+        })
+    })
     assert test == dict1
     dict1 = AttributeDict(init_dict1)
     del dict2.dct1.test
@@ -62,7 +64,14 @@ def test_nested_dict_delete(init_dict1):
     """Test nested dict delete."""
     dict1 = AttributeDict(init_dict1)
     delete_keys_from_dict(dict1, 'dct1.test')
-    test = AttributeDict({'dct1': AttributeDict({'test2': 'string1', 'test3': [1.0], 'test4': AttributeDict({'key1': 'value1'})})})
+    test = AttributeDict({
+        'dct1':
+        AttributeDict({
+            'test2': 'string1',
+            'test3': [1.0],
+            'test4': AttributeDict({'key1': 'value1'})
+        })
+    })
     assert test == dict1
     delete_keys_from_dict(dict1, 'dct1.test5')
     assert test == dict1
