@@ -343,7 +343,9 @@ class VaspNEBWorkChain(BaseRestartWorkChain):
         if ibrion != 3:
             self.report('WARNING: IBRION should be set to 3 for VTST runs, proceed with caution.')
         elif potim != 0:
-            self.report('WARNING: Using VTST optimisors with IBRION=3, but POTIM is not set to zero, proceed with caution.')
+            self.report(
+                'WARNING: Using VTST optimisors with IBRION=3, but POTIM is not set to zero, proceed with caution.'
+            )
         if iopt == 0:
             self.report('WARNING: IOPT not set.')
 
@@ -372,8 +374,11 @@ class VaspNEBWorkChain(BaseRestartWorkChain):
             norm_disp = np.linalg.norm(disp, axis=1)
             sort_idx = np.argsort(norm_disp)
             if norm_disp[sort_idx[-1]] > self._norm_disp_threshold:
-                raise InputValidationError('Large displacement detected for atom {} at frame {} - please check the inputs images'.format(
-                    sort_idx[-1], iframe + 1))
+                raise InputValidationError(
+                    'Large displacement detected for atom {} at frame {} - please check the inputs images'.format(
+                        sort_idx[-1], iframe + 1
+                    )
+                )
             last_frame = frame
 
     def _setup_vasp_inputs(self):
@@ -467,13 +472,15 @@ class VaspNEBWorkChain(BaseRestartWorkChain):
         # Verify and set potentials (potcar)
         if not self.inputs.potential_family.value:
             self.report(  # pylint: disable=not-callable
-                'An empty string for the potential family name was detected.')
+                'An empty string for the potential family name was detected.'
+            )
             return self.exit_codes.ERROR_NO_POTENTIAL_FAMILY_NAME  # pylint: disable=no-member
         try:
             self.ctx.inputs.potential = get_data_class('vasp.potcar').get_potcars_from_structure(
                 structure=self.inputs.initial_structure,
                 family_name=self.inputs.potential_family.value,
-                mapping=self.inputs.potential_mapping.get_dict())
+                mapping=self.inputs.potential_mapping.get_dict()
+            )
         except ValueError as err:
             return compose_exit_code(self.exit_codes.ERROR_POTENTIAL_VALUE_ERROR.status, str(err))  # pylint: disable=no-member
         except NotExistent as err:
@@ -493,8 +500,10 @@ class VaspNEBWorkChain(BaseRestartWorkChain):
         # for the outcome of the work chain and so have marked it as `is_finished=True`.
         max_iterations = self.inputs.max_iterations.value  # type: ignore[union-attr]
         if not self.ctx.is_finished and self.ctx.iteration >= max_iterations:
-            self.report(f'reached the maximum number of iterations {max_iterations}: '
-                        f'last ran {self.ctx.process_name}<{node.pk}>')
+            self.report(
+                f'reached the maximum number of iterations {max_iterations}: '
+                f'last ran {self.ctx.process_name}<{node.pk}>'
+            )
             return self.exit_codes.ERROR_MAXIMUM_ITERATIONS_EXCEEDED  # pylint: disable=no-member
 
         self.report(f'work chain completed after {self.ctx.iteration} iterations')
@@ -597,7 +606,14 @@ def get_ldau_keys(structure, mapping, utype=2, jmapping=None, felec=False):
 
     if count > 0:
         # Only enable U is there is any non-zero value
-        output = {'ldauu': ldauu, 'ldauj': ldauj, 'ldautype': utype, 'lmaxmix': 6 if felec else 4, 'ldaul': ldaul, 'ldau': True}
+        output = {
+            'ldauu': ldauu,
+            'ldauj': ldauj,
+            'ldautype': utype,
+            'lmaxmix': 6 if felec else 4,
+            'ldaul': ldaul,
+            'ldau': True
+        }
     else:
         output = {}
     return output
