@@ -38,7 +38,9 @@ class EosWorkChain(WorkChain):
     def define(cls, spec):
         super(EosWorkChain, cls).define(spec)
         spec.expose_inputs(cls._next_workchain, exclude=['structure'])
-        spec.input_namespace('structures', valid_type=DataFactory('structure'), dynamic=True, help='a dictionary of structures to use')
+        spec.input_namespace(
+            'structures', valid_type=DataFactory('structure'), dynamic=True, help='a dictionary of structures to use'
+        )
         spec.exit_code(0, 'NO_ERROR', message='the sun is shining')
         spec.exit_code(420, 'ERROR_NO_CALLED_WORKCHAIN', message='no called workchain detected')
         spec.exit_code(500, 'ERROR_UNKNOWN', message='unknown error detected in the eos workchain')
@@ -75,7 +77,7 @@ class EosWorkChain(WorkChain):
         self.ctx.is_finished = False
 
         # Define an interation index
-        self.ctx.interation = 0
+        self.ctx.iteration = 0
 
         # Define the context inputs
         self.ctx.inputs = AttributeDict()
@@ -134,9 +136,8 @@ class EosWorkChain(WorkChain):
         """
         inputs = self.ctx.inputs
         running = self.submit(self._next_workchain, **inputs)
-
         self.report(f'launching {self._next_workchain.__name__}<{running.pk}> iteration #{self.ctx.iteration}')
-        return self.to_context(workchains=append_(running))
+        self.to_context(workchains=append_(running))
 
     def verify_next_workchain(self):
         """Correct for unexpected behavior."""
@@ -155,10 +156,14 @@ class EosWorkChain(WorkChain):
             self.ctx.exit_code = self.exit_codes.NO_ERROR  # pylint: disable=no-member
         else:
             self.ctx.exit_code = compose_exit_code(next_workchain_exit_status, next_workchain_exit_message)
-            self.report('The called {}<{}> returned a non-zero exit status. '
-                        'The exit status {} is inherited'.format(workchain.__class__.__name__, workchain.pk, self.ctx.exit_code))
+            self.report(
+                'The called {}<{}> returned a non-zero exit status. '
+                'The exit status {} is inherited'.format(
+                    workchain.__class__.__name__, workchain.pk, self.ctx.exit_code
+                )
+            )
 
-        # Stop further executions of workchains if there are no more structure
+        # Stop further execution of workchains if there are no more structure
         # entries in the structures dictionary
         if not self.ctx.structures:
             self.ctx.is_finished = True
