@@ -31,36 +31,36 @@ def main(code_string, incar, kmesh, structure, potential_family, potential_mappi
     kpoints_data = DataFactory('array.kpoints')
 
     # Then, we set the workchain you would like to call
-    workchain = WorkflowFactory('vasp.verify')
+    workchain = WorkflowFactory('vasp.vasp')
 
     # And finally, we declare the options, settings and input containers
     settings = AttributeDict()
     inputs = AttributeDict()
 
-    # organize settings
+    # Organize settings
     settings.parser_settings = {'add_dos': True}
 
-    # set inputs for the following WorkChain execution
-    # set code
+    # Set inputs for the following WorkChain execution
+    # Code
     inputs.code = Code.get_from_string(code_string)
-    # set structure
+    # Structure
     inputs.structure = structure
-    # set k-points grid density
+    # k-points grid density
     kpoints = kpoints_data()
     kpoints.set_kpoints_mesh(kmesh)
     inputs.kpoints = kpoints
-    # set parameters
+    # Parameters
     inputs.parameters = dict_data(dict=incar)
-    # set potentials and their mapping
+    # Potential family and their mapping between element and potential type to use
     inputs.potential_family = Str(potential_family)
     inputs.potential_mapping = dict_data(dict=potential_mapping)
-    # set options
+    # Options
     inputs.options = dict_data(dict=options)
-    # set settings
+    # Settings
     inputs.settings = dict_data(dict=settings)
-    # set workchain related inputs, in this case, give more explicit output to report
+    # Workchain related inputs, in this case, give more explicit output to report
     inputs.verbose = Bool(True)
-    # submit the requested workchain with the supplied inputs
+    # Submit the workchain with the set inputs
     submit(workchain, **inputs)
 
 
@@ -70,7 +70,7 @@ if __name__ == '__main__':
 
     # INCAR equivalent
     # Set input parameters
-    INCAR = {'encut': 240, 'ismear': -5, 'lorbit': 11}
+    INCAR = {'incar': {'encut': 240, 'ismear': -5, 'lorbit': 11}}
 
     # KPOINTS equivalent
     # Set kpoint mesh
@@ -79,7 +79,7 @@ if __name__ == '__main__':
     # POTCAR equivalent
     # Potential_family is chosen among the list given by
     # 'verdi data vasp-potcar listfamilies'
-    POTENTIAL_FAMILY = 'pbe'
+    POTENTIAL_FAMILY = 'PBE.54'
     # The potential mapping selects which potential to use, here we use the standard
     # for silicon, this could for instance be {'Si': 'Si_GW'} to use the GW ready
     # potential instead
@@ -93,10 +93,10 @@ if __name__ == '__main__':
     OPTIONS = AttributeDict()
     OPTIONS.account = ''
     OPTIONS.qos = ''
-    OPTIONS.resources = {'num_machines': 1, 'num_mpiprocs_per_machine': 16}
+    OPTIONS.resources = {'num_machines': 1, 'num_mpiprocs_per_machine': 8}
     OPTIONS.queue_name = ''
     OPTIONS.max_wallclock_seconds = 3600
-    OPTIONS.max_memory_kb = 1024000
+    OPTIONS.max_memory_kb = 2000000
 
     # POSCAR equivalent
     # Set the silicon structure
