@@ -36,7 +36,7 @@ def get_structure(alat):
     structure_data = DataFactory('structure')
     lattice = np.array([[.5, .5, 0], [0, .5, .5], [.5, 0, .5]]) * alat
     structure = structure_data(cell=lattice)
-    for pos_direct in ([[0.0, 0.0, 0.0]]):
+    for pos_direct in [[0.0, 0.0, 0.0]]:
         pos_cartesian = np.dot(pos_direct, lattice)
         structure.append_atom(position=pos_cartesian, symbols='Si')
     return structure
@@ -51,47 +51,44 @@ def main(code_string, incar, kmesh, structure, potential_family, potential_mappi
     kpoints_data = DataFactory('array.kpoints')
 
     # Then, we set the workchain you would like to call
-    workchain = WorkflowFactory('vasp.verify')
+    workchain = WorkflowFactory('vasp.vasp')
 
     # And finally, we declare the options, settings and input containers
     settings = AttributeDict()
     inputs = AttributeDict()
 
-    # Organize settings
-    settings.parser_settings = {'output_params': ['total_energies', 'maximum_force']}
-
     # Set inputs for the following WorkChain execution
-    # Set code
+    # Code
     inputs.code = Code.get_from_string(code_string)
-    # Set structure
+    # Structure
     inputs.structure = structure
-    # Set k-points grid density
+    # k-points grid density
     kpoints = kpoints_data()
     kpoints.set_kpoints_mesh(kmesh)
     inputs.kpoints = kpoints
-    # Set parameters
+    # Parameters
     inputs.parameters = dict_data(dict=incar)
-    # Set potentials and their mapping
+    # Potential family and their mapping between element and potential type to use
     inputs.potential_family = Str(potential_family)
     inputs.potential_mapping = dict_data(dict=potential_mapping)
-    # Set options
+    # Options
     inputs.options = dict_data(dict=options)
-    # Set settings
+    # Settings
     inputs.settings = dict_data(dict=settings)
-    # Set workchain related inputs, in this case, give more explicit output to report
+    # Workchain related inputs, in this case, give more explicit output to report
     inputs.verbose = Bool(True)
-    # Submit the requested workchain with the supplied inputs
+    # Submit the workchain with the set inputs
     results = run(workchain, **inputs)
     return results
 
 
 if __name__ == '__main__':
     # Code_string is chosen among the list given by 'verdi code list'
-    CODE_STRING = 'vasp@mycluster'
+    CODE_STRING = 'VASP/6.3.2-gompi-2021b-std-wannier90-libxc-hdf5-beef-d7238be44ec2ed23315a16cc1549a1e3@betzy'
 
     # INCAR equivalent
     # Set input parameters
-    INCAR = {'istart': 0, 'icharg': 2, 'encut': 240, 'ismear': 0, 'sigma': 0.1}
+    INCAR = {'incar': {'istart': 0, 'icharg': 2, 'encut': 240, 'ismear': 0, 'sigma': 0.1}}
 
     # KPOINTS equivalent
     # Set kpoint mesh
@@ -100,7 +97,7 @@ if __name__ == '__main__':
     # POTCAR equivalent
     # Potential_family is chosen among the list given by
     # 'verdi data vasp-potcar listfamilies'
-    POTENTIAL_FAMILY = 'pbe'
+    POTENTIAL_FAMILY = 'PBE'
     # The potential mapping selects which potential to use, here we use the standard
     # for silicon, this could for instance be {'Si': 'Si_GW'} to use the GW ready
     # potential instead
@@ -112,12 +109,12 @@ if __name__ == '__main__':
     # AttributeDict is just a special dictionary with the extra benefit that
     # you can set and get the key contents with mydict.mykey, instead of mydict['mykey']
     OPTIONS = AttributeDict()
-    OPTIONS.account = ''
-    OPTIONS.qos = ''
-    OPTIONS.resources = {'num_machines': 1, 'num_mpiprocs_per_machine': 16}
+    OPTIONS.account = 'nn9997k'
+    OPTIONS.qos = 'devel'
+    OPTIONS.resources = {'num_machines': 1, 'num_mpiprocs_per_machine': 8}
     OPTIONS.queue_name = ''
     OPTIONS.max_wallclock_seconds = 3600
-    OPTIONS.max_memory_kb = 1024000
+    OPTIONS.max_memory_kb = 2000000
 
     EOS = []
     # Iterate over each lattice constant and pass it explicitly
@@ -126,10 +123,11 @@ if __name__ == '__main__':
         # Set the silicon structure
         STRUCTURE = get_structure(lattice_constant)
 
-        output = main(CODE_STRING, INCAR, KMESH, STRUCTURE, POTENTIAL_FAMILY, POTENTIAL_MAPPING, OPTIONS)
+        results = main(CODE_STRING, INCAR, KMESH, STRUCTURE, POTENTIAL_FAMILY, POTENTIAL_MAPPING, OPTIONS)
+
         # The output are stored as AiiDA datatypes, which is Dict in this case. To obtain a regular
         # dictionary, we use get_dict
-        misc = output['misc'].get_dict()
+        misc = results['misc'].get_dict()
         EOS.append([lattice_constant, misc['total_energies']['energy_extrapolated']])
     # Write volume and total energies to file
     with open('eos', 'w', encoding='utf8') as file_object:
