@@ -36,8 +36,7 @@ def get_structure():
     alat = 3.9
     lattice = np.array([[.5, .5, 0], [0, .5, .5], [.5, 0, .5]]) * alat
     structure = structure_data(cell=lattice)
-    positions = [[0.0, 0.0, 0.0]]
-    for pos_direct in positions:
+    for pos_direct in [[0.0, 0.0, 0.0]]:
         pos_cartesian = np.dot(pos_direct, lattice)
         structure.append_atom(position=pos_cartesian, symbols='Si')
     return structure
@@ -52,36 +51,33 @@ def main(code_string, incar, kmesh, structure, potential_family, potential_mappi
     kpoints_data = DataFactory('array.kpoints')
 
     # Then, we set the workchain you would like to call
-    workchain = WorkflowFactory('vasp.verify')
+    workchain = WorkflowFactory('vasp.vasp')
 
     # And finally, we declare the options, settings and input containers
     settings = AttributeDict()
     inputs = AttributeDict()
 
-    # organize settings
-    settings.parser_settings = {'output_params': ['total_energies', 'maximum_force']}
-
-    # set inputs for the following WorkChain execution
-    # set code
+    # Set inputs for the following WorkChain execution
+    # Code
     inputs.code = Code.get_from_string(code_string)
-    # set structure
+    # Structure
     inputs.structure = structure
-    # set k-points grid density
+    # k-points grid density
     kpoints = kpoints_data()
     kpoints.set_kpoints_mesh(kmesh)
     inputs.kpoints = kpoints
-    # set parameters
+    # Parameters
     inputs.parameters = dict_data(dict=incar)
-    # set potentials and their mapping
+    # Potential family and their mapping between element and potential type to use
     inputs.potential_family = Str(potential_family)
     inputs.potential_mapping = dict_data(dict=potential_mapping)
-    # set options
+    # Options
     inputs.options = dict_data(dict=options)
-    # set settings
+    # Settings
     inputs.settings = dict_data(dict=settings)
-    # set workchain related inputs, in this case, give more explicit output to report
+    # Workchain related inputs, in this case, give more explicit output to report
     inputs.verbose = Bool(True)
-    # submit the requested workchain with the supplied inputs
+    # Submit the workchain with the set inputs
     submit(workchain, **inputs)
 
 
@@ -91,7 +87,7 @@ if __name__ == '__main__':
 
     # INCAR equivalent
     # Set input parameters
-    INCAR = {'istart': 0, 'icharg': 2, 'encut': 240, 'ismear': 0, 'sigma': 0.1}
+    INCAR = {'incar': {'istart': 0, 'icharg': 2, 'encut': 240, 'ismear': 0, 'sigma': 0.1}}
 
     # KPOINTS equivalent
     # Set kpoint mesh
@@ -100,13 +96,13 @@ if __name__ == '__main__':
     # POTCAR equivalent
     # Potential_family is chosen among the list given by
     # 'verdi data vasp-potcar listfamilies'
-    POTENTIAL_FAMILY = 'pbe'
+    POTENTIAL_FAMILY = 'PBE.54'
     # The potential mapping selects which potential to use, here we use the standard
     # for silicon, this could for instance be {'Si': 'Si_GW'} to use the GW ready
     # potential instead
     POTENTIAL_MAPPING = {'Si': 'Si'}
 
-    # jobfile equivalent
+    # Jobfile equivalent
     # In options, we typically set scheduler options.
     # See https://aiida.readthedocs.io/projects/aiida-core/en/latest/scheduler/index.html
     # AttributeDict is just a special dictionary with the extra benefit that
@@ -114,10 +110,10 @@ if __name__ == '__main__':
     OPTIONS = AttributeDict()
     OPTIONS.account = ''
     OPTIONS.qos = ''
-    OPTIONS.resources = {'num_machines': 1, 'num_mpiprocs_per_machine': 16}
+    OPTIONS.resources = {'num_machines': 1, 'num_mpiprocs_per_machine': 1}
     OPTIONS.queue_name = ''
     OPTIONS.max_wallclock_seconds = 3600
-    OPTIONS.max_memory_kb = 1024000
+    OPTIONS.max_memory_kb = 2000000
 
     # POSCAR equivalent
     # Set the silicon structure
